tesseract-engine/requirements.txt
numpy==1.15.4
pillow>=6.2.0
pytesseract==0.2.6
fastapi==0.59.0
uvicorn==0.11.5